+++ doc/howto/write_a_view/solution_view.cpp
@@ -1,10 +1,11 @@
 //![iterator]
 #include <iostream>
+#include <seqan3/std/ranges>
 #include <vector>
+
 #include <seqan3/alphabet/nucleotide/all.hpp>
 #include <seqan3/core/debug_stream.hpp>
-#include <seqan3/range/detail/inherited_iterator_base.hpp>
-#include <seqan3/std/ranges>
+#include <seqan3/core/range/detail/inherited_iterator_base.hpp>
 
 using seqan3::operator""_dna5;
