+++ include/seqan3/alphabet/views/translate.hpp
@@ -23,7 +23,6 @@
 #include <seqan3/core/add_enum_bitwise_operators.hpp>
 #include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/container/small_string.hpp>
-#include <seqan3/range/detail/random_access_iterator.hpp>
 #include <seqan3/range/views/detail.hpp>
 #include <seqan3/range/container/concept.hpp>
 #include <seqan3/utility/views/deep.hpp>
