+++ include/seqan3/io/sam_file/detail/format_sam_base.hpp
@@ -25,6 +25,7 @@
 #include <seqan3/core/debug_stream/detail/to_string.hpp>
 #include <seqan3/core/detail/debug_stream_range.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
+#include <seqan3/core/range/detail/misc.hpp>
 #include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/io/detail/ignore_output_iterator.hpp>
 #include <seqan3/io/detail/istreambuf_view.hpp>
@@ -37,7 +38,6 @@
 #include <seqan3/io/sam_file/output_options.hpp>
 #include <seqan3/io/sam_file/sam_tag_dictionary.hpp>
 #include <seqan3/io/sequence_file/output_options.hpp>
-#include <seqan3/range/detail/misc.hpp>
 #include <seqan3/range/views/take_until.hpp>
 #include <seqan3/range/views/to.hpp>
 #include <seqan3/utility/detail/exposition_only_concept.hpp>
