+++ include/seqan3/range/container/bitcompressed_vector.hpp
@@ -23,7 +23,7 @@
 #include <seqan3/alphabet/views/to_char.hpp>
 #include <seqan3/alphabet/views/to_rank.hpp>
 #include <seqan3/core/concept/cereal.hpp>
-#include <seqan3/range/detail/random_access_iterator.hpp>
+#include <seqan3/core/range/detail/random_access_iterator.hpp>
 #include <seqan3/utility/math.hpp>
 #include <seqan3/utility/views/convert.hpp>
