+++ include/seqan3/range/container/concatenated_sequences.hpp
@@ -19,7 +19,6 @@
 
 #include <seqan3/core/concept/cereal.hpp>
 #include <seqan3/range/container/concept.hpp>
-#include <seqan3/range/detail/random_access_iterator.hpp>
 #include <seqan3/range/views/as_const.hpp>
 #include <seqan3/utility/views/repeat_n.hpp>
 #include <seqan3/utility/views/slice.hpp>
