+++ include/seqan3/range/decorator/gap_decorator.hpp
@@ -25,8 +25,6 @@
 #include <seqan3/alphabet/gap/gap.hpp>
 #include <seqan3/alphabet/gap/gapped.hpp>
 #include <seqan3/range/container/concept.hpp>
-#include <seqan3/range/detail/inherited_iterator_base.hpp>
-#include <seqan3/range/detail/random_access_iterator.hpp>
 #include <seqan3/range/views/type_reduce.hpp>
 
 namespace seqan3
@@ -139,7 +137,7 @@ public:
      */
     using unaligned_sequence_type = inner_type;
 
-    //!\deprecated Please use seqan3::gap_decorator::unaligned_sequence_type instead.   
+    //!\deprecated Please use seqan3::gap_decorator::unaligned_sequence_type instead.
     using unaligned_seq_type SEQAN3_DEPRECATED_310 = unaligned_sequence_type;
 
     /*!\name Constructors, destructor and assignment
