+++ include/seqan3/range/views/take_until.hpp
@@ -19,12 +19,12 @@
 #include <seqan3/std/type_traits>
 
 #include <seqan3/core/detail/iterator_traits.hpp>
+#include <seqan3/core/range/detail/inherited_iterator_base.hpp>
 #include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/core/semiregular_box.hpp>
 #include <seqan3/io/exception.hpp>
 #include <seqan3/range/concept.hpp>
 #include <seqan3/range/views/detail.hpp>
-#include <seqan3/range/detail/inherited_iterator_base.hpp>
 #include <seqan3/range/container/concept.hpp>
 #include <seqan3/utility/type_traits/detail/transformation_trait_or.hpp>
