+++ include/seqan3/utility/views/enforce_random_access.hpp
@@ -15,8 +15,8 @@
 #include <seqan3/std/iterator>
 #include <seqan3/std/ranges>
 
+#include <seqan3/core/range/detail/inherited_iterator_base.hpp>
 #include <seqan3/range/views/detail.hpp>
-#include <seqan3/range/detail/inherited_iterator_base.hpp>
 #include <seqan3/range/concept.hpp>
 
 namespace seqan3::detail
