+++ include/seqan3/utility/views/repeat.hpp
@@ -16,8 +16,8 @@
 #include <seqan3/std/ranges>
 
 #include <seqan3/core/detail/iterator_traits.hpp>
+#include <seqan3/core/range/detail/random_access_iterator.hpp>
 #include <seqan3/core/range/type_traits.hpp>
-#include <seqan3/range/detail/random_access_iterator.hpp>
 
 namespace seqan3::detail
 {
