+++ test/unit/core/range/range_iterator_test.cpp
@@ -11,8 +11,8 @@
 
 #include <gtest/gtest.h>
 
+#include <seqan3/core/range/detail/random_access_iterator.hpp>
 #include <seqan3/core/range/type_traits.hpp>
-#include <seqan3/range/detail/random_access_iterator.hpp>
 #include <seqan3/range/views/take_exactly.hpp>
 #include <seqan3/utility/detail/type_name_as_string.hpp>
 #include <seqan3/utility/type_list/traits.hpp>
